SVGRenderingAccuracyTest now saves candidate reference image in 'candidate-ref' sub-directory (along with candidate-variation and accepted-variation) sub-directories when the reference image cannot be opened. This makes the task of reviewing new test images easier. PR: Obtained from: Submitted by: Reviewed by:
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@200339 13f79535-47bb-0310-9956-ffa450edef68

+++ test-sources/org/apache/batik/test/svg/SVGRenderingAccuracyTest.java
@@ -249,6 +249,13 @@ public class SVGRenderingAccuracyTest extends AbstractTest {
     protected File saveVariation;
 
     /**
+     * The File where the candidate reference 
+     * should be saved if there is not candidate reference
+     * or if it cannot be opened.
+     */
+    protected File candidateReference;
+
+    /**
      * Temporary directory
      */
     protected static File tempDirectory;
@@ -353,6 +360,17 @@ public class SVGRenderingAccuracyTest extends AbstractTest {
      */
     public void setVariationURL(String variationURL){
         this.variationURL = resolveURL(variationURL);
+    }
+
+    /**
+     * See {@link #candidateReference}
+     */
+    public void setCandidateReference(File candidateReference){
+        this.candidateReference = candidateReference;
+    }
+
+    public File getCandidateReference(){
+        return candidateReference;
     }
 
     /**
@@ -479,7 +497,18 @@ public class SVGRenderingAccuracyTest extends AbstractTest {
                                                  new Object[]{refImgURL.toString(), 
                                                               e.getMessage()})) });
             report.setPassed(false);
-            tmpFile.delete();
+            // Try and save tmp file as a candidate variation
+            boolean deleteTmp = true;
+            if (candidateReference != null){
+                if (candidateReference.exists()){
+                    candidateReference.delete();
+                }
+                deleteTmp = tmpFile.renameTo(candidateReference);
+            }
+
+            if (deleteTmp){
+                tmpFile.delete();
+            }
             return report;
         }
